pyproject.toml
[project]
name = "gov-doc-praser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "pydantic-ai>=0.0.24",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.0.1",
]
